requirements.txt
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.7
dulwich==0.21.7
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pastel==0.2.1
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poethepoet==0.24.4
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
protobuf==4.25.1
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
temporalio==1.4.0
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.29
types-protobuf==4.24.0.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
xattr==0.10.1
zipp==3.17.0